pyproject.toml
[tool]
[tool.poetry]
name = "kube-downscaler"
version = "2020.1.1dev0"
description = "Scale down Kubernetes deployments after work hours"
authors = ["Henning Jacobs <henning@jacobs1.de>"]

[tool.poetry.dependencies]
python = ">=3.7"
pykube-ng = "*"
pytz = "*"
humanize = "^2.4.0"
python-dateutil = "^2.8.1"
google-api-python-client = "^1.8.2"
tzlocal = "^2.0.0"

[tool.poetry.dev-dependencies]
coverage = "*"
coveralls = "*"
flake8 = "*"
pytest = "*"
pytest-cov = "*"
black = "^19.10b0"
mypy = "^0.761"
pre-commit = "^1.21.0"
